# Conflicts:
#	Discord.Net
#	scripts/Latest.bat
#	scripts/NadekoAutoRun.bat
#	scripts/Stable.bat
#	src/NadekoBot/NadekoBot.cs
#	src/NadekoBot/Resources/CommandStrings.Designer.cs
#	src/NadekoBot/Resources/CommandStrings.resx
#	src/NadekoBot/Resources/ResponseStrings.de-DE.resx
#	src/NadekoBot/Resources/ResponseStrings.en-US.resx
#	src/NadekoBot/Resources/ResponseStrings.es-ES.resx
#	src/NadekoBot/Resources/ResponseStrings.fr-FR.resx
#	src/NadekoBot/Resources/ResponseStrings.he-IL.resx
#	src/NadekoBot/Resources/ResponseStrings.it-IT.resx
#	src/NadekoBot/Resources/ResponseStrings.ja-JP.resx
#	src/NadekoBot/Resources/ResponseStrings.ko-KR.resx
#	src/NadekoBot/Resources/ResponseStrings.nb-NO.resx
#	src/NadekoBot/Resources/ResponseStrings.nl-NL.resx
#	src/NadekoBot/Resources/ResponseStrings.pl-PL.resx
#	src/NadekoBot/Resources/ResponseStrings.pt-BR.resx
#	src/NadekoBot/Resources/ResponseStrings.ru-RU.resx
#	src/NadekoBot/Resources/ResponseStrings.sr-cyrl-rs.resx
#	src/NadekoBot/Resources/ResponseStrings.sv-SE.resx
#	src/NadekoBot/Resources/ResponseStrings.tr-TR.resx
#	src/NadekoBot/Resources/ResponseStrings.zh-CN.resx
#	src/NadekoBot/Resources/ResponseStrings.zh-TW.resx
#	src/NadekoBot/Services/Database/IUnitOfWork.cs
#	src/NadekoBot/Services/Database/UnitOfWork.cs
#	src/NadekoBot/project.json

+++ src/NadekoBot/Modules/Administration/Commands/SelfAssignedRolesCommand.cs
@@ -18,14 +18,20 @@ namespace NadekoBot.Modules.Administration
         [Group]
         public class SelfAssignedRolesCommands : NadekoSubmodule
         {
-            
+            private readonly DbService _db;
+
+            public SelfAssignedRolesCommands(DbService db)
+            {
+                _db = db;
+            }
+
             [NadekoCommand, Usage, Description, Aliases]
             [RequireContext(ContextType.Guild)]
             [RequireUserPermission(GuildPermission.ManageMessages)]
             public async Task AdSarm()
             {
                 bool newval;
-                using (var uow = DbHandler.UnitOfWork())
+                using (var uow = _db.UnitOfWork)
                 {
                     var config = uow.GuildConfigs.For(Context.Guild.Id, set => set);
                     newval = config.AutoDeleteSelfAssignedRoleMessages = !config.AutoDeleteSelfAssignedRoleMessages;
@@ -43,9 +49,13 @@ namespace NadekoBot.Modules.Administration
             {
                 IEnumerable<SelfAssignedRole> roles;
 
+                var guser = (IGuildUser)Context.User;
+                if (Context.User.Id != guser.Guild.OwnerId && guser.GetRoles().Max(x => x.Position) <= role.Position)
+                    return;
+
                 string msg;
                 var error = false;
-                using (var uow = DbHandler.UnitOfWork())
+                using (var uow = _db.UnitOfWork)
                 {
                     roles = uow.SelfAssignedRoles.GetFromGuild(Context.Guild.Id);
                     if (roles.Any(s => s.RoleId == role.Id && s.GuildId == role.Guild.Id))
@@ -75,8 +85,12 @@ namespace NadekoBot.Modules.Administration
             [RequireUserPermission(GuildPermission.ManageRoles)]
             public async Task Rsar([Remainder] IRole role)
             {
+                var guser = (IGuildUser)Context.User;
+                if (Context.User.Id != guser.Guild.OwnerId && guser.GetRoles().Max(x => x.Position) <= role.Position)
+                    return;
+
                 bool success;
-                using (var uow = DbHandler.UnitOfWork())
+                using (var uow = _db.UnitOfWork)
                 {
                     success = uow.SelfAssignedRoles.DeleteByGuildAndRoleId(role.Guild.Id, role.Id);
                     await uow.CompleteAsync();
@@ -95,24 +109,24 @@ namespace NadekoBot.Modules.Administration
             {
                 var toRemove = new ConcurrentHashSet<SelfAssignedRole>();
                 var removeMsg = new StringBuilder();
-                var msg = new StringBuilder();
+                var roles = new List<string>();
                 var roleCnt = 0;
-                using (var uow = DbHandler.UnitOfWork())
+                using (var uow = _db.UnitOfWork)
                 {
                     var roleModels = uow.SelfAssignedRoles.GetFromGuild(Context.Guild.Id).ToList();
-                    roleCnt = roleModels.Count;
-                    msg.AppendLine();
                     
                     foreach (var roleModel in roleModels)
                     {
                         var role = Context.Guild.Roles.FirstOrDefault(r => r.Id == roleModel.RoleId);
                         if (role == null)
                         {
+                            toRemove.Add(roleModel);
                             uow.SelfAssignedRoles.Remove(roleModel);
                         }
                         else
                         {
-                            msg.Append($"**{role.Name}**, ");
+                            roles.Add(Format.Bold(role.Name));
+                            roleCnt++;
                         }
                     }
                     foreach (var role in toRemove)
@@ -121,7 +135,7 @@ namespace NadekoBot.Modules.Administration
                     }
                     await uow.CompleteAsync();
                 }
-                await Context.Channel.SendConfirmAsync(GetText("self_assign_list", roleCnt), msg + "\n\n" + removeMsg).ConfigureAwait(false);
+                await Context.Channel.SendConfirmAsync(GetText("self_assign_list", roleCnt), "\n" + string.Join(", ", roles) + "\n\n" + removeMsg).ConfigureAwait(false);
             }
 
             [NadekoCommand, Usage, Description, Aliases]
@@ -130,7 +144,7 @@ namespace NadekoBot.Modules.Administration
             public async Task Tesar()
             {
                 bool areExclusive;
-                using (var uow = DbHandler.UnitOfWork())
+                using (var uow = _db.UnitOfWork)
                 {
                     var config = uow.GuildConfigs.For(Context.Guild.Id, set => set);
 
@@ -151,7 +165,7 @@ namespace NadekoBot.Modules.Administration
 
                 GuildConfig conf;
                 SelfAssignedRole[] roles;
-                using (var uow = DbHandler.UnitOfWork())
+                using (var uow = _db.UnitOfWork)
                 {
                     conf = uow.GuildConfigs.For(Context.Guild.Id, set => set);
                     roles = uow.SelfAssignedRoles.GetFromGuild(Context.Guild.Id).ToArray();
@@ -177,7 +191,7 @@ namespace NadekoBot.Modules.Administration
                         var sameRole = Context.Guild.GetRole(sameRoleId);
                         if (sameRole != null)
                         {
-                            await guildUser.RemoveRolesAsync(sameRole).ConfigureAwait(false);
+                            await guildUser.RemoveRoleAsync(sameRole).ConfigureAwait(false);
                             await Task.Delay(500).ConfigureAwait(false);
                         }
                         //await ReplyErrorLocalized("self_assign_already_excl", Format.Bold(sameRole?.Name)).ConfigureAwait(false);
@@ -186,7 +200,7 @@ namespace NadekoBot.Modules.Administration
                 }
                 try
                 {
-                    await guildUser.AddRolesAsync(role).ConfigureAwait(false);
+                    await guildUser.AddRoleAsync(role).ConfigureAwait(false);
                 }
                 catch (Exception ex)
                 {
@@ -211,7 +225,7 @@ namespace NadekoBot.Modules.Administration
 
                 bool autoDeleteSelfAssignedRoleMessages;
                 IEnumerable<SelfAssignedRole> roles;
-                using (var uow = DbHandler.UnitOfWork())
+                using (var uow = _db.UnitOfWork)
                 {
                     autoDeleteSelfAssignedRoleMessages = uow.GuildConfigs.For(Context.Guild.Id, set => set).AutoDeleteSelfAssignedRoleMessages;
                     roles = uow.SelfAssignedRoles.GetFromGuild(Context.Guild.Id);
@@ -228,7 +242,7 @@ namespace NadekoBot.Modules.Administration
                 }
                 try
                 {
-                    await guildUser.RemoveRolesAsync(role).ConfigureAwait(false);
+                    await guildUser.RemoveRoleAsync(role).ConfigureAwait(false);
                 }
                 catch (Exception)
                 {
